Improve generator code path (#38)
